Update Hyperscale Data Center in Wyoming_V2
 This WyomingCredits contract builds on the previous version by adding the ability to mint USAT tokens when environmental credits are redeemed.

USAT stands for "Utility Settlement Coin - Asset Token". It is a stablecoin pegged to $1 USD.

When a user redeems carbon credits, an equivalent value of USATs are minted based on the set carbon credit pricing of $20 per tonne.

For example, if a user redeems 100 carbon credits, that is worth 100 * $20 = $2000. So 2000/1 = 2000 USAT tokens would be minted and sent to the user.

Similarly for water and ESRIN credits, the contract mints USATs worth the dollar value of the credits based on their set prices.

This provides a few benefits:

    Users get liquidity from redeeming their credits, in the form of USAT tokens which are pegged to USD.

    The total USATs minted represents the dollar value of redeemed credits, tracking the revenue generated.

    USATs can be easily transferred, spent or traded since they are ERC20 tokens on blockchain.

    The backing of USATs by redeemed credits gives the stablecoin intrinsic value.

This upgrade allows converting illiquid environmental credits into liquid stablecoin assets by minting USATs during redemption. The totals track revenue, and users get usable tokens representing the dollar value of their credits.
Emiliano Solazzi